requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.9
numpy==1.26.0
soundfile==0.12.1
librosa==0.10.1
torch==2.2.0
demucs==4.0.0
openai-whisper==20240207
noisereduce==2.0.1
webrtcvad==2.0.10
requests==2.32.0
python-dotenv==1.1.1
